i45: Add useAppLaunch custom hook
