+++ src/library/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useLocation } from './useLocation';
 export { default as useMap } from './useMap';
+export { default as useAppLaunch } from './useAppLaunch';
